Add files via upload
This was mine first project when I was learning HTML and CSS I got an idea to create a website to test my skills. It turned out to make an website for a gym that it can host to attract more and more people towards their gym. I just made this for testing my skills there was no idea to turn it out into a business and there would not be any in the future.